feat(core): land useDragGesture via @/ import + local vue-lynx worklet patch
Re-lands the shared useDragGesture composable (reverted in #38) now that the
real blocker is understood: vue-lynx's MT worklet loader only follows relative
imports, so the @/-aliased composable's worklets never registered on MT
(`bind of undefined`). Not a vue-lynx-vs-composable limitation — an import-path
packaging gap.

- Re-add shared/gesture/useDragGesture.ts; SwiperRoot consumes it via @/ import
- patches/vue-lynx@0.4.0.patch: pnpm patch widening worklet-loader-mt's
  extractLocalImports to also follow @/ aliases (REMOVE WHEN UPSTREAM FIXED;
  undo steps in patches/README.md)
- docs/upstream/: drafted vue-lynx issue (root cause + tiered fix)

Verified: MT-bundle audit unresolved worklets 7 -> 0. Device tap-test pending.
Relates to #37, #38.
