+++ pnpm-workspace.yaml
@@ -16,3 +16,5 @@ allowBuilds:
   esbuild: true
   sharp: true
   vue-demi: false
+patchedDependencies:
+  vue-lynx@0.4.0: patches/vue-lynx@0.4.0.patch
